+++ .github/workflows/test-lz-endpoint.yml
@@ -1,0 +1,76 @@
+name: "@hyperbridge/lz-endpoint"
+
+on:
+    push:
+        branches:
+            - main
+        paths:
+            - "sdk/packages/lz-endpoint/**"
+            - ".github/workflows/test-lz-endpoint.yml"
+    pull_request_target:
+        branches:
+            - main
+        types: [opened, synchronize, reopened, ready_for_review]
+        paths:
+            - "sdk/packages/lz-endpoint/**"
+            - ".github/workflows/test-lz-endpoint.yml"
+
+concurrency:
+    group: ${{ github.ref }}-lz-endpoint
+    cancel-in-progress: true
+
+env:
+    FOUNDRY_PROFILE: default
+
+jobs:
+    test:
+        name: Build & Test
+        if: github.event_name != 'pull_request_target' || github.event.pull_request.draft == false
+        runs-on: ubuntu-latest
+        defaults:
+            run:
+                working-directory: sdk
+        steps:
+            - uses: actions/checkout@v3
+
+            - name: Install Foundry
+              uses: foundry-rs/foundry-toolchain@v1
+              with:
+                  version: nightly
+
+            - name: Setup Node.js
+              uses: actions/setup-node@v3
+              with:
+                  node-version: "18"
+
+            - uses: pnpm/action-setup@v4
+              with:
+                  version: 10
+                  run_install: false
+
+            - name: Get pnpm store directory
+              id: pnpm-cache
+              shell: bash
+              run: |
+                  echo "STORE_PATH=$(pnpm store path)" >> $GITHUB_OUTPUT
+
+            - name: Setup pnpm cache
+              uses: actions/cache@v3
+              with:
+                  path: ${{ steps.pnpm-cache.outputs.STORE_PATH }}
+                  key: ${{ runner.os }}-pnpm-store-${{ hashFiles('**/pnpm-lock.yaml') }}
+                  restore-keys: |
+                      ${{ runner.os }}-pnpm-store-
+
+            - name: Install dependencies
+              run: pnpm install
+
+            - name: Build
+              working-directory: sdk/packages/lz-endpoint
+              run: forge build
+
+            - name: Test
+              working-directory: sdk/packages/lz-endpoint
+              env:
+                  MAINNET_FORK_URL: ${{ secrets.MAINNET_FORK_URL }}
+              run: forge test -vvv
