+++ modules/pallets/hyper-fungible-token/src/impls.rs
@@ -19,7 +19,7 @@ use alloc::string::ToString;
 use polkadot_sdk::*;
 use sp_core::U256;
 
-use crate::{Config, Pallet};
+use crate::{Config, Pallet, PALLET_ID};
 
 impl<T: Config> Pallet<T> {
 	/// Returns the pallet's custodial account for holding native assets
@@ -28,20 +28,27 @@ impl<T: Config> Pallet<T> {
 		use sp_runtime::traits::AccountIdConversion;
 		PalletId(*b"hft__acc").into_account_truncating()
 	}
+
+	/// Returns true if the given module ID matches this pallet's well-known ID.
+	/// Used by the runtime's IsmpRouter to route messages to this pallet.
+	pub fn is_module(id: &[u8]) -> bool {
+		id == PALLET_ID.to_bytes()
+	}
 }
 
-/// Converts an ERC20 U256 amount to a local u128 balance
+/// Converts an ERC20 U256 amount to a local balance type
 ///
-/// Divides by 10^(erc_decimals - local_decimals) to scale down from ERC20 precision
-pub fn convert_to_balance(
+/// Divides by 10^(erc_decimals - local_decimals) to scale down from ERC20 precision.
+/// The target type must implement `FromStr`.
+pub fn convert_to_balance<B: core::str::FromStr>(
 	value: U256,
 	erc_decimals: u8,
 	local_decimals: u8,
-) -> Result<u128, anyhow::Error> {
+) -> Result<B, B::Err> {
 	let dec_str = (value /
 		U256::from(10u128.pow(erc_decimals.saturating_sub(local_decimals) as u32)))
 	.to_string();
-	dec_str.parse().map_err(|e| anyhow::anyhow!("{e:?}"))
+	dec_str.parse::<B>()
 }
 
 /// Converts a local u128 balance to an ERC20 U256 amount
