+++ sdk/packages/core/contracts/apps/WrappedHyperFungibleToken.sol
@@ -30,7 +30,7 @@ import {HyperFungibleToken} from "./HyperFungibleToken.sol";
 /**
  * @title WrappedHyperFungibleToken
  * @author Polytope Labs (hello@polytope.technology)
- * @notice Abstract base contract for wrapping existing ERC20 tokens for cross-chain transfers.
+ * @notice Cross-chain wrapper for existing ERC20 tokens.
  * Locks the underlying token on the source chain and mints/unlocks on the destination chain.
  *
  * @dev Inherits HyperApp for cross-chain message handling and Ownable for configuration.
@@ -44,7 +44,7 @@ import {HyperFungibleToken} from "./HyperFungibleToken.sol";
  * Supports optional calldata execution on the destination chain via CallDispatcher,
  * enabling composable cross-chain interactions (e.g., transfer-and-swap).
  */
-abstract contract WrappedHyperFungibleToken is HyperApp, Ownable, Pausable {
+contract WrappedHyperFungibleToken is HyperApp, Ownable, Pausable {
     /**
      * @title WrappedConfigOptions
      * @notice Configuration parameters for WrappedHyperFungibleToken
@@ -90,10 +90,10 @@ abstract contract WrappedHyperFungibleToken is HyperApp, Ownable, Pausable {
     bool internal _isWeth;
 
     /**
-     * @notice Maps chain identifiers to the address of the wrapper contract on that chain.
-     * A zero address means the chain is not supported.
+     * @notice Maps chain identifiers to the module ID of the peer on that chain.
+     * An empty value means the chain is not supported.
      */
-    mapping(bytes => address) internal _supportedChains;
+    mapping(bytes => bytes) internal _supportedChains;
 
     /**
      * @notice Emitted when tokens are locked and a cross-chain transfer is dispatched
@@ -103,7 +103,7 @@ abstract contract WrappedHyperFungibleToken is HyperApp, Ownable, Pausable {
      * @param amount The amount of tokens sent
      * @param commitment The ISMP request commitment hash for tracking
      */
-    event Sent(bytes from, bytes to, bytes dest, uint256 amount, bytes32 commitment);
+    event Sent(address from, bytes to, string dest, uint256 amount, bytes32 commitment);
 
     /**
      * @notice Emitted when tokens are unlocked from an incoming cross-chain transfer
@@ -112,7 +112,7 @@ abstract contract WrappedHyperFungibleToken is HyperApp, Ownable, Pausable {
      * @param source The source chain identifier
      * @param amount The amount of tokens unlocked
      */
-    event Received(bytes from, bytes to, bytes source, uint256 amount);
+    event Received(bytes from, address to, string source, uint256 amount);
 
     /**
      * @notice Emitted when tokens are refunded after a cross-chain transfer timeout
@@ -159,11 +159,11 @@ abstract contract WrappedHyperFungibleToken is HyperApp, Ownable, Pausable {
     }
 
     /**
-     * @notice Returns the wrapper contract address for a given chain
+     * @notice Returns the module ID of the peer on a given chain
      * @param chainId The chain identifier
-     * @return The address of the wrapper contract on the specified chain
+     * @return The module ID of the peer on the specified chain
      */
-    function supportedChain(bytes calldata chainId) public view returns (address) {
+    function supportedChain(bytes calldata chainId) public view returns (bytes memory) {
         return _supportedChains[chainId];
     }
 
@@ -183,10 +183,10 @@ abstract contract WrappedHyperFungibleToken is HyperApp, Ownable, Pausable {
      * @notice Registers a supported chain and its corresponding wrapper contract address
      * @dev Only callable by the contract owner
      * @param chainId The chain identifier (e.g., StateMachine.evm(1))
-     * @param contractAddress The address of the wrapper contract on the specified chain
+     * @param moduleId The module ID of the peer on the specified chain
      */
-    function addChain(bytes calldata chainId, address contractAddress) external onlyOwner {
-        _supportedChains[chainId] = contractAddress;
+    function addChain(bytes calldata chainId, bytes calldata moduleId) external onlyOwner {
+        _supportedChains[chainId] = moduleId;
     }
 
     /**
@@ -226,8 +226,8 @@ abstract contract WrappedHyperFungibleToken is HyperApp, Ownable, Pausable {
      * @param params The send parameters including destination, recipient, amount, and optional calldata
      */
     function send(HyperFungibleToken.SendParams calldata params) external payable whenNotPaused {
-        address dest = _supportedChains[params.dest];
-        if (dest == address(0)) revert UnsupportedChain();
+        bytes memory dest = _supportedChains[params.dest];
+        if (dest.length == 0) revert UnsupportedChain();
 
         uint256 msgValue = msg.value;
         if (_isWeth) {
@@ -247,7 +247,7 @@ abstract contract WrappedHyperFungibleToken is HyperApp, Ownable, Pausable {
 
         DispatchPost memory request = DispatchPost({
             dest: params.dest,
-            to: abi.encodePacked(dest),
+            to: dest,
             body: body,
             timeout: params.timeout,
             fee: params.relayerFee,
@@ -261,7 +261,7 @@ abstract contract WrappedHyperFungibleToken is HyperApp, Ownable, Pausable {
             commitment = dispatchWithFeeToken(request, msg.sender);
         }
 
-        emit Sent(from, params.to, params.dest, params.amount, commitment);
+        emit Sent(msg.sender, params.to, string(params.dest), params.amount, commitment);
     }
 
     /**
@@ -274,9 +274,9 @@ abstract contract WrappedHyperFungibleToken is HyperApp, Ownable, Pausable {
     function onAccept(IncomingPostRequest calldata incoming) external override onlyHost whenNotPaused {
         PostRequest calldata request = incoming.request;
 
-        address expectedSource = _supportedChains[request.source];
-        if (expectedSource == address(0)) revert UnsupportedChain();
-        if (address(bytes20(request.from[:20])) != expectedSource) revert UnauthorizedSource();
+        bytes memory expectedSource = _supportedChains[request.source];
+        if (expectedSource.length == 0) revert UnsupportedChain();
+        if (keccak256(request.from) != keccak256(expectedSource)) revert UnauthorizedSource();
 
         HyperFungibleToken.Message memory message = abi.decode(request.body, (HyperFungibleToken.Message));
         address beneficiary = _toAddr(message.to);
@@ -286,7 +286,7 @@ abstract contract WrappedHyperFungibleToken is HyperApp, Ownable, Pausable {
             ICallDispatcher(_dispatcher).dispatch(message.data);
         }
 
-        emit Received(message.from, abi.encodePacked(beneficiary), request.source, message.amount);
+        emit Received(message.from, beneficiary, string(request.source), message.amount);
     }
 
     /**
